[FEATURE] Allow Storages outside the webroot
Storages can be marked as not public or be outside of webroot.
There are no direct links to the files posible.

This change adds a eID implementation to create urls to these files.
It also implements a hook in het eID dispatcher that makes is possible
to add custom security/access checks when delivering files.

Releases: 6.2
Resolves: #36420
Reviewed-on: https://review.typo3.org/27760
Tested-by: Frans Saris
Tested-by: Fabien Udriot
Tested-by: Steffen Ritter

+++ typo3/sysext/core/Classes/Resource/AbstractRepository.php
@@ -204,7 +204,7 @@ abstract class AbstractRepository implements \TYPO3\CMS\Extbase\Persistence\Repo
 	 * @return string the additional where clause, something like " AND deleted=0 AND hidden=0"
 	 */
 	protected function getWhereClauseForEnabledFields() {
-		if ($this->getEnvironmentMode() === 'FE') {
+		if ($this->getEnvironmentMode() === 'FE' && $GLOBALS['TSFE']->sys_page) {
 			// frontend context
 			$whereClause = $GLOBALS['TSFE']->sys_page->enableFields($this->table);
 			$whereClause .= $GLOBALS['TSFE']->sys_page->deleteClause($this->table);
